+++ Changelog.md
@@ -1,5 +1,5 @@
 ## 0.0.3
-*INCOMPLETE*
++ Refactored `Dtype.Undefined` enumeration member to `Dtype.Null`.
 
 ## 0.0.2
 + Added `FunctionUnity.ToAudioClip` extension method for converting a Function `Value` to an `AudioClip`.